+++ DragNWashCoop/CoopPlugin.cs
@@ -9,7 +9,7 @@ using UnityEngine.SceneManagement;
 
 namespace DragNWashCoop;
 
-[BepInPlugin(Guid, "Co-Washers", "0.5.1")]
+[BepInPlugin(Guid, "Co-Washers", "0.5.2")]
 public sealed class CoopPlugin : BaseUnityPlugin
 {
     public const string Guid = "trixiescience.dragnwashcoop";